client/server/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
python-dotenv==1.0.0
Werkzeug>=2.3.3